Update CutsceneMacroMakerV2.js
few minor bugfixes, moved the dialog for the CMM to the right hand side of the screen to try to not obstruct user workspace obnoxiously. this introduced a new bug where when you first try to move the dialog it seems offset from the mouse for some reason, also it doesnt remember the last position so it felt like a better idea for it to have this issue off to the side rather than the middle of the screen.